fix: format auth folders
