+++ .env.example
@@ -5,6 +5,7 @@ JWT_SECRET_KEY=secret123123
 JWT_SECRET_REFRESH_KEY=secret123123
 TOKEN_EXPIRE_TIME=1h
 TOKEN_REFRESH_EXPIRE_TIME=24h
+REQUEST_USER_KEY = 'user';
 
 JWT_SECRET=YOUR_SECRET_KEY_HERE
 JWT_TOKEN_AUDIENCE=localhost:4000
