+++ src/auth/auth.module.ts
@@ -1,9 +1,4 @@
-import {
-  MiddlewareConsumer,
-  Module,
-  NestModule,
-  RequestMethod,
-} from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthController } from './auth.controller';
@@ -11,17 +6,17 @@ import { jwtConstants } from './constants';
 import { UserModule } from 'src/user/user.module';
 import { UserService } from 'src/user/user.service';
 import { PrismaService } from 'src/database/prisma.service';
-import { HashingService } from 'src/iam/hashing/hashing.service';
-import { BcryptService } from 'src/iam/hashing/bcrypt.service';
-import jwtConfig from 'src/iam/config/jwt.config';
 import { ConfigModule } from '@nestjs/config';
 import { APP_FILTER, APP_GUARD, APP_INTERCEPTOR } from '@nestjs/core';
-import { AccessTokenGuard } from 'src/iam/auth/guards/access-token.guard';
-import { AuthenticationGuard } from 'src/iam/auth/guards/authentication.guard';
+import { AccessTokenGuard } from 'src/auth/guards/access-token.guard';
+import { AuthenticationGuard } from 'src/auth/guards/authentication.guard';
 
 import { CustomLogger } from 'src/custom-logger/custom-logger.service';
 import { LoggingInterceptor } from 'src/custom-logger/logging-interceptor';
 import { HttpExceptionFilter } from 'src/exception-filter/exception-filter';
+import jwtConfig from './config/jwt.config';
+import { HashingService } from './hashing/hashing.service';
+import { BcryptService } from './hashing/bcrypt.service';
 
 @Module({
   imports: [
