+++ src/auth/config/jwt.config.ts
