+++ src/auth/decorators/auth.decorator.ts
@@ -1,7 +1,18 @@
 import { SetMetadata } from '@nestjs/common';
 import { AuthType } from '../enums/auth-type.enum';
+import { ExecutionContext, createParamDecorator } from '@nestjs/common';
+import { ActiveUserData } from '../interfaces/active-user-data.interface';
 
 export const AUTH_TYPE_KEY = 'authType';
 
 export const Auth = (...authTypes: AuthType[]) =>
   SetMetadata(AUTH_TYPE_KEY, authTypes);
+
+export const ActiveUser = createParamDecorator(
+  (field: keyof ActiveUserData | undefined, ctx: ExecutionContext) => {
+    const request = ctx.switchToHttp().getRequest();
+    const user: ActiveUserData | undefined =
+      request[process.env.REQUEST_USER_KEY];
+    return field ? user?.[field] : user;
+  },
+);
