+++ src/auth/guards/authentication.guard.spec.ts
@@ -1,0 +1,7 @@
+// import { AuthenticationGuard } from './authentication.guard';
+
+// describe('AuthenticationGuard', () => {
+//   it('should be defined', () => {
+//     expect().toBeDefined();
+//   });
+// });
