--- src/iam/auth/guards/authentication.guard.spec.ts
@@ -1,7 +1,0 @@
-import { AuthenticationGuard } from './authentication.guard';
-
-describe('AuthenticationGuard', () => {
-  it('should be defined', () => {
-    expect(new AuthenticationGuard()).toBeDefined();
-  });
-});
