--- src/iam/decorators/active-user.decorator.ts
@@ -1,11 +1,0 @@
-import { ExecutionContext, createParamDecorator } from '@nestjs/common';
-import { REQUEST_USER_KEY } from '../iam.constants';
-import { ActiveUserData } from '../interfaces/active-user-datea.interface';
-
-export const ActiveUser = createParamDecorator(
-  (field: keyof ActiveUserData | undefined, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    const user: ActiveUserData | undefined = request[REQUEST_USER_KEY];
-    return field ? user?.[field] : user;
-  },
-);
