--- src/iam/iam.constants.ts
@@ -1,1 +1,0 @@
-export const REQUEST_USER_KEY = 'user';
